Moved subscriptions into  SqlQueryService
